Rename gem, lib and database to departure
