+++ lib/departure/configuration.rb
