+++ lib/departure/null_logger.rb
