+++ lib/departure/user_options.rb
