+++ spec/spec_helper.rb
@@ -8,7 +8,7 @@ Bundler.require(:default, :development)
 require './configuration'
 require './test_database'
 
-require 'percona_migrator'
+require 'departure'
 require 'lhm'
 
 require 'support/matchers/have_column'
@@ -26,7 +26,7 @@ ActiveRecord::Base.establish_connection(
   host: 'localhost',
   username: db_config['username'],
   password: db_config['password'],
-  database: 'percona_migrator_test'
+  database: db_config['database']
 )
 
 MIGRATION_FIXTURES = File.expand_path('../fixtures/migrate/', __FILE__)
@@ -39,7 +39,7 @@ RSpec.configure do |config|
   ActiveRecord::Migration.verbose = false
 
   # Needs an empty block to initialize the config with the default values
-  PerconaMigrator.configure do |config|
+  PerconaMigrator.configure do |_config|
   end
 
   config.before(:each) do |example|
